chore: cargo fmt, drop unnecessary u64 casts, log live-demo-readyup
Pre-existing lints caught by the mw-cp checkpoint:

- src/runner.rs: drop two `as u64` casts on `result.instructions_run`
  (already `u64`); clippy -D warnings was failing on them.
- src/runner.rs + tests/demos.rs: apply cargo fmt (wrap assert_eq!
  args; strip trailing blank line).

Also records the live-demo-readyup saga under `[Unreleased]` in
CHANGES.md: new demos (higher-order-lists, function-keyword,
function-pattern-args, when-guards), GH Pages deploy workflow, and
the README live-demo link + screenshot restructure.

+++ src/runner.rs
@@ -192,8 +192,8 @@ impl Session {
 
             let chunk = INNER_BATCH.min(budget - total);
             let result = self.emu.run_batch(chunk);
-            self.instructions += result.instructions_run as u64;
-            total += result.instructions_run as u64;
+            self.instructions += result.instructions_run;
+            total += result.instructions_run;
 
             match result.reason {
                 StopReason::Halted => {
@@ -364,7 +364,8 @@ mod tests {
         );
         let cleaned = s.clean_output();
         assert_eq!(
-            cleaned, "42",
+            cleaned,
+            "42",
             "expected exactly '42' after stripping echo, got: {cleaned:?} (raw: {:?})",
             s.output()
         );
